backend/vehicle_service_management/requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers